# docs/api/openapi.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "Koc.Dwos.Api",
        "version": "1.0"
    },
    "paths": {
...
        "/api/movement-plans": {
            "get": {
                "tags": [
                    "MovementPlans"
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            },
            "post": {
                "tags": [
                    "MovementPlans"
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/JsonNode"
                            }
                        },
                        "text/json": {
                            "schema": {
                                "$ref": "#/components/schemas/JsonNode"
                            }
                        },
                        "application/*+json": {
                            "schema": {
                                "$ref": "#/components/schemas/JsonNode"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/movement-plans/{date}": {
            "get": {
                "tags": [
                    "MovementPlans"
                ],
                "parameters": [
                    {
                        "name": "date",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
        "/api/org/entities": {
            "get": {
                "tags": [
                    "Org"
                ],
                "parameters": [
                    {
                        "name": "type",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
        "/api/reports": {
            "get": {
                "tags": [
                    "Reports"
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            },
            "post": {
                "tags": [
                    "Reports"
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/JsonNode"
                            }
                        },
                        "text/json": {
                            "schema": {
                                "$ref": "#/components/schemas/JsonNode"
                            }
                        },
                        "application/*+json": {
                            "schema": {
                                "$ref": "#/components/schemas/JsonNode"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/reports/{vesselCode}/{date}": {
            "get": {
                "tags": [
                    "Reports"
                ],
                "parameters": [
                    {
                        "name": "vesselCode",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "date",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "JsonNode": {
                "type": "object",
                "properties": {
                    "Options": {
                        "$ref": "#/components/schemas/JsonNodeOptions"
                    },
                    "Parent": {
                        "$ref": "#/components/schemas/JsonNode"
                    },
                    "Root": {
                        "$ref": "#/components/schemas/JsonNode"
                    }
                },
                "additionalProperties": false
            },
            "JsonNodeOptions": {
                "type": "object",
                "properties": {
                    "PropertyNameCaseInsensitive": {
                        "type": "boolean"
                    }
                },
                "additionalProperties": false
            }
        }
    }
}